Add Deck.gl PathLayer to a mapgl::maplibre() or mapgl::mapboxgl() map using blazing fast nanoarrow::write_nanoarrow() data transfer.

Add Deck.gl PathLayer to a mapgl::maplibre() or mapgl::mapboxgl() map using blazing fast nanoarrow::write_nanoarrow() data transfer.

Usage

addGeoArrowPathLayer(
  map,
  data = NULL,
  url = NULL,
  layer_id = "path",
  geom_column_name = "geometry",
  popup = NULL,
  tooltip = NULL,
  render_options = renderOptions(),
  data_accessors = dataAccessors(),
  popup_options = popupOptions(),
  tooltip_options = tooltipOptions(),
  ...
)

Arguments

map

the mapgl::maplibre() or mapgl::mapboxgl() map to add the layer to.

data

a sf (MULTI)LINESTRING object.

url

a URL to a remotely hosted geoarrow or geoparquet file to be added to the map. Ignored if data is supplied.

layer_id

the layer id.

geom_column_name

the name of the geometry column of the sf object. It is inferred automatically if only one is present.

popup

should a popup be contructed? If TRUE, will create a popup fromm all available attributes of the feature. Can also be a character vector of column names, on which case the popup will include only those columns. If a single character is supplied, then this will be shown for all features. If NULL (deafult) or FALSE, no popup will be shown.

tooltip

should a tooltip be contructed? If TRUE, will create a tooltip fromm all available attributes of the feature. Can also be a character vector of column names, on which case the tooltip will include only those columns. If a single character is supplied, then this will be shown for all features. If NULL (deafult) or FALSE, no tooltip will be shown.

render_options

a list of renderOptions

data_accessors

a list of dataAccessors

popup_options

a list of popupOptions

tooltip_options

a list of tooltipOptions

...

currently not used.